jwt fix

+++ backend/controllers/Messages.js
@@ -3,9 +3,23 @@
 var utils = require('../utils/writer.js');
 var Messages = require('../service/MessagesService');
 
-module.exports.deleteMessage = function deleteMessage(req, res, next, chatId, messageId) {
-    const userId = req.user.id;
+const jwt = require('jsonwebtoken');
 
+function getUserIdFromToken(req, res) {
+    try {
+        const token = req.headers.authorization?.split(' ')[1];
+        if (!token) throw new Error('No token');
+        const user = jwt.verify(token, process.env.JWT_SECRET);
+        return user.id || user.userId;
+    } catch (e) {
+        utils.writeJson(res, { message: 'User is not authenticated or token is invalid', error: e.message, headers: req.headers }, 401);
+        return null;
+    }
+}
+
+module.exports.deleteMessage = function deleteMessage(req, res, next, chatId, messageId) {
+    const userId = getUserIdFromToken(req, res);
+    if (!userId) return;
     Messages.deleteMessage(chatId, messageId, userId)
         .then(function(response) {
             utils.writeJson(res, response);
@@ -16,8 +30,8 @@ module.exports.deleteMessage = function deleteMessage(req, res, next, chatId, me
 };
 
 module.exports.editMessage = function editMessage(req, res, next, body, chatId, messageId) {
-    const userId = req.user.id;
-
+    const userId = getUserIdFromToken(req, res);
+    if (!userId) return;
     Messages.editMessage(chatId, messageId, userId, body.content)
         .then(function(response) {
             utils.writeJson(res, response);
@@ -28,8 +42,8 @@ module.exports.editMessage = function editMessage(req, res, next, body, chatId, 
 };
 
 module.exports.getMessages = function getMessages(req, res, next, chatId) {
-    const userId = req.user.id;
-
+    const userId = getUserIdFromToken(req, res);
+    if (!userId) return;
     Messages.getMessages(chatId, userId)
         .then(function(response) {
             utils.writeJson(res, response);
@@ -40,8 +54,8 @@ module.exports.getMessages = function getMessages(req, res, next, chatId) {
 };
 
 module.exports.sendMessage = function sendMessage(req, res, next, body, chatId) {
-    const userId = req.user.id;
-
+    const userId = getUserIdFromToken(req, res);
+    if (!userId) return;
     Messages.sendMessage(body, chatId, userId)
         .then(function(response) {
             utils.writeJson(res, response);
